+++ .vs/PingViewerApp/v17/DocumentLayout.json
@@ -1,0 +1,109 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "C:\\Users\\renyr\\source\\repos\\PingApp\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|c:\\users\\renyr\\source\\repos\\pingapp\\bussines\\services\\pingmonitor.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|solutionrelative:bussines\\services\\pingmonitor.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|c:\\users\\renyr\\source\\repos\\pingapp\\views\\dashboardwindow\\dashboardwindow.axaml.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|solutionrelative:views\\dashboardwindow\\dashboardwindow.axaml.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|c:\\users\\renyr\\source\\repos\\pingapp\\viewmodels\\pingresultsviewmodel.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|solutionrelative:viewmodels\\pingresultsviewmodel.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|c:\\users\\renyr\\source\\repos\\pingapp\\views\\dashboardwindow\\dashboardwindow.axaml||{FA3CD31E-987B-443A-9B81-186104E8DAC1}",
+      "RelativeMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|solutionrelative:views\\dashboardwindow\\dashboardwindow.axaml||{FA3CD31E-987B-443A-9B81-186104E8DAC1}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|c:\\users\\renyr\\source\\repos\\pingapp\\bussines\\services\\pingrequester.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{8318F8CD-3D14-09BB-6D63-B68E19F45962}|PingViewerApp.csproj|solutionrelative:bussines\\services\\pingrequester.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    }
+  ],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": [
+        {
+          "DockedWidth": 228,
+          "SelectedChildIndex": 2,
+          "Children": [
+            {
+              "$type": "Bookmark",
+              "Name": "ST:0:0:{1c4feeaa-4718-4aa9-859d-94ce25d182ba}"
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 1,
+              "Title": "DashboardWindow.axaml.cs",
+              "DocumentMoniker": "C:\\Users\\renyr\\source\\repos\\PingApp\\Views\\DashboardWindow\\DashboardWindow.axaml.cs",
+              "RelativeDocumentMoniker": "Views\\DashboardWindow\\DashboardWindow.axaml.cs",
+              "ToolTip": "C:\\Users\\renyr\\source\\repos\\PingApp\\Views\\DashboardWindow\\DashboardWindow.axaml.cs",
+              "RelativeToolTip": "Views\\DashboardWindow\\DashboardWindow.axaml.cs",
+              "ViewState": "AgIAAAoAAAAAAAAAAAAQwDMAAAAqAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-08-06T22:23:04.836Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "PingMonitor.cs",
+              "DocumentMoniker": "C:\\Users\\renyr\\source\\repos\\PingApp\\Bussines\\Services\\PingMonitor.cs",
+              "RelativeDocumentMoniker": "Bussines\\Services\\PingMonitor.cs",
+              "ToolTip": "C:\\Users\\renyr\\source\\repos\\PingApp\\Bussines\\Services\\PingMonitor.cs",
+              "RelativeToolTip": "Bussines\\Services\\PingMonitor.cs",
+              "ViewState": "AgIAAEEAAAAAAAAAAAAUwE4AAAAkAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-08-06T22:46:20.237Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 4,
+              "Title": "PingRequester.cs",
+              "DocumentMoniker": "C:\\Users\\renyr\\source\\repos\\PingApp\\Bussines\\Services\\PingRequester.cs",
+              "RelativeDocumentMoniker": "Bussines\\Services\\PingRequester.cs",
+              "ToolTip": "C:\\Users\\renyr\\source\\repos\\PingApp\\Bussines\\Services\\PingRequester.cs",
+              "RelativeToolTip": "Bussines\\Services\\PingRequester.cs",
+              "ViewState": "AgIAABoAAAAAAAAAAAAswBoAAABkAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-08-06T22:46:41.017Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 3,
+              "Title": "DashboardWindow.axaml",
+              "DocumentMoniker": "C:\\Users\\renyr\\source\\repos\\PingApp\\Views\\DashboardWindow\\DashboardWindow.axaml",
+              "RelativeDocumentMoniker": "Views\\DashboardWindow\\DashboardWindow.axaml",
+              "ToolTip": "C:\\Users\\renyr\\source\\repos\\PingApp\\Views\\DashboardWindow\\DashboardWindow.axaml",
+              "RelativeToolTip": "Views\\DashboardWindow\\DashboardWindow.axaml",
+              "ViewState": "AgIAAB4AAAAAAAAAAAAqwCUAAAAtAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001001|",
+              "WhenOpened": "2025-08-05T21:41:25.26Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 2,
+              "Title": "PingResultsViewModel.cs",
+              "DocumentMoniker": "C:\\Users\\renyr\\source\\repos\\PingApp\\ViewModels\\PingResultsViewModel.cs",
+              "RelativeDocumentMoniker": "ViewModels\\PingResultsViewModel.cs",
+              "ToolTip": "C:\\Users\\renyr\\source\\repos\\PingApp\\ViewModels\\PingResultsViewModel.cs",
+              "RelativeToolTip": "ViewModels\\PingResultsViewModel.cs",
+              "ViewState": "AgIAAEMAAAAAAAAAAAAawFQAAAAnAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-08-08T06:20:22.333Z",
+              "EditorCaption": ""
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
