+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "SelectedNode": "\\PingViewerApp.sln",
+  "PreviewInSolutionExplorer": false
+}
