--- App.axaml.cs
@@ -1,35 +1,0 @@
-using Avalonia;
-using Avalonia.Controls.ApplicationLifetimes;
-using Avalonia.Data.Core.Plugins;
-using Avalonia.Markup.Xaml;
-using Microsoft.Extensions.DependencyInjection;
-using PingViewerApp.Utils.Extensions;
-using System;
-
-namespace PingViewerApp;
-
-public partial class App : Application
-{
-    public static IServiceProvider Services { get; private set; }
-    public override void Initialize()
-    {
-        AvaloniaXamlLoader.Load(this);
-    }
-
-    public override void OnFrameworkInitializationCompleted()
-    {
-        var collection = new ServiceCollection();
-        collection.AddViewModels();
-        collection.AddViews();
-        collection.AddCommonServices();
-
-        Services = collection.BuildServiceProvider();
-
-        if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
-        {
-            desktop.MainWindow = new MainWindow();
-        }
-
-        base.OnFrameworkInitializationCompleted();
-    }
-}
