+++ App/Assets/Icons/icon.png
