+++ App/Bussines/Entities/PingItem.cs
