+++ App/Bussines/Services/PingRequester.cs
