+++ App/Models/PingResult.cs
