+++ App/ViewModels/PingResultsViewModel.cs
