+++ App/Views/DashboardWindow/DashboardWindow.axaml.cs
